backend/pyproject.toml
[project]
name = "decision-ledger-backend"
version = "0.1.0"
description = "Decision Ledger backend — ingest, extract, resolve, detect, narrate."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "supabase>=2.4",
    "anthropic>=0.40",
    "google-genai>=0.3",
    "numpy>=1.26",
    "pydantic>=2.6",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
package = false
